[CrashFix, DirectX, bin] Add files (read description)
The "CrashFix" folder contains a significant number of binaries such as
Visual C++ runtimes, DirectX files, and other binaries used in the
installation and subsequent customization of Fallout Fixt. Due to space
concerns and the nature of GitHub, these files are not uploaded here.
However, they can be provided upon request.