[Installers] Remove savegames-backup and Registry
No longer will the installers back up savegames. This is because XCOPY
was causing issues on some systems. I would like to reimplement
automatic backups of savegames at some point, but I didn't like my
current implementation of it anyway.
Also removed registry entries because when we switch to it installing in
the subfolder \Fallout Fixt\, I don't want it to mess with vanilla
registry entries.
Lastly, updated my notes-to-self at the top :)